requirements.txt
# Core dependencies
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
python-dateutil>=2.8.2

# AI and NLP
google-generativeai>=0.3.0
langdetect>=1.0.9

# Scheduling
APScheduler>=3.10.4

# Web scraping
newspaper3k>=0.2.8
lxml>=4.9.3

# Data storage
